+++ sdk/servicebus/service-bus/src/receivers/receiver.ts
@@ -37,6 +37,8 @@ export interface Receiver<ReceivedMessageT> {
    * Streams messages to message handlers.
    * @param handlers A handler that gets called for messages and errors.
    * @param options Options for subscribe.
+   * @returns An object that can be closed, sending any remaining messages to `handlers` and
+   * stopping new messages from arriving.
    */
   subscribe(
     handlers: MessageHandlers<ReceivedMessageT>,
@@ -403,7 +405,7 @@ export class ReceiverImpl<ReceivedMessageT extends ReceivedMessage | ReceivedMes
 
     return {
       close: async (): Promise<void> => {
-        return this._context.streamingReceiver?.stopReceivingMessages();
+        return this._context.streamingReceiver?.receiverHelper.stopReceivingMessages();
       }
     };
   }