+++ sdk/servicebus/service-bus/src/receivers/sessionReceiver.ts
@@ -453,8 +453,9 @@ export class SessionReceiverImpl<ReceivedMessageT extends ReceivedMessage | Rece
     );
 
     return {
-      // TODO: coming in a future PR.
-      close: async (): Promise<void> => {}
+      close: async (): Promise<void> => {
+        return this._messageSession?.receiverHelper.stopReceivingMessages();
+      }
     };
   }